+++ packages/global-styles-ui/src/screen-header.tsx
@@ -4,9 +4,7 @@
 import {
 	__experimentalHStack as HStack,
 	__experimentalVStack as VStack,
-	__experimentalSpacer as Spacer,
 	__experimentalHeading as Heading,
-	__experimentalView as View,
 	Navigator,
 } from '@wordpress/components';
 import { isRTL, __ } from '@wordpress/i18n';
@@ -24,28 +22,22 @@ export function ScreenHeader( {
 	onBack,
 }: ScreenHeaderProps ) {
 	return (
-		<VStack spacing={ 0 }>
-			<View>
-				<Spacer marginBottom={ 0 } paddingX={ 4 } paddingY={ 3 }>
-					<HStack spacing={ 2 }>
-						<Navigator.BackButton
-							icon={ isRTL() ? chevronRight : chevronLeft }
-							size="small"
-							label={ __( 'Back' ) }
-							onClick={ onBack }
-						/>
-						<Spacer>
-							<Heading
-								className="global-styles-ui-header"
-								level={ 2 }
-								size={ 13 }
-							>
-								{ title }
-							</Heading>
-						</Spacer>
-					</HStack>
-				</Spacer>
-			</View>
+		<VStack spacing={ 0 } className="global-styles-ui-screen-header">
+			<HStack spacing={ 2 } justify="flex-start">
+				<Navigator.BackButton
+					icon={ isRTL() ? chevronRight : chevronLeft }
+					size="small"
+					label={ __( 'Back' ) }
+					onClick={ onBack }
+				/>
+				<Heading
+					className="global-styles-ui-header"
+					level={ 2 }
+					size={ 13 }
+				>
+					{ title }
+				</Heading>
+			</HStack>
 			{ description && (
 				<p className="global-styles-ui-header__description">
 					{ description }